+++ fs/fjs/module.f.ts
@@ -3,36 +3,35 @@
  *
  * @module
  */
-import { fromIo, type Io } from '../io/module.f.ts'
 import { compile } from '../djs/module.f.ts'
 import { main as testMain } from '../dev/tf/module.f.ts'
 import { main as casMain } from '../cas/module.f.ts'
-import type { NodeProgram } from '../types/effects/node/module.f.ts'
+import { import_, error, type NodeProgram } from '../types/effects/node/module.f.ts'
+import { pure } from '../types/effects/module.f.ts'
 
-export const main = async(io: Io): Promise<number> => {
-    const { error } = io.console
-    const [command, ...rest] = io.process.argv.slice(2)
-    const eRun = fromIo(io)
+export const main: NodeProgram = options => {
+    const [command, ...rest] = options.args
     switch (command) {
         case 'test':
         case 't':
-            return testMain(io)
+            return testMain({ ...options, args: rest })
         case 'compile':
         case 'c':
-            return compile(io)(rest)
+            return compile(rest)
         case 'cas':
         case 's':
-            return eRun(casMain(rest))
+            return casMain(rest)
         case 'run':
-        case 'r':
+        case 'r': {
             const [file, ...args] = rest
-            const m = await io.asyncImport(file)
-            return eRun((m.default as NodeProgram)(args))
+            return import_(file).step(([s, v]) => {
+                if (s === 'error') { throw v }
+                return (v.default as NodeProgram)({ ...options, args })
+            })
+        }
         case undefined:
-            error('Error: command is required')
-            return Promise.resolve(1)
+            return error('Error: command is required').step(() => pure(1))
         default:
-            error(`Error: Unknown command "${command}"`)
-            return Promise.resolve(1)
+            return error(`Error: Unknown command "${command}"`).step(() => pure(1))
     }
 }